# openapi.yaml
openapi: 3.1.0
info:
  title: Gozne
  version: 0.1.0-dev.4
  description:
    Alpha EVM/SIWS authentication and signed simulation controls. API and login
    require the same HTTPS origin. Authentication returns 503 without a policy.
paths:
...
  /v1/auth/control:
    get:
      summary: Read the current application control overview
      description:
        Up to 100 recent actions. Readers see actions from their current session
        only. Administrators also receive up to 100 invitations and active
        sessions. All authorized users receive up to 20 simulated deployment
        receipts. Counters should not be treated as lifetime totals.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlOverview'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/control/invitations:
    post:
      summary: Grant temporary reader access to an unconfigured wallet
      description:
        Requires admin and reader in the current application. Static wallet
        entries cannot be invited. Only applications requiring no role other
        than reader support invitations. Validity starts immediately; acceptance
        is the first successful wallet login.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationCreated'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                network:
                  type: string
                  enum:
                    - evm
                    - solana
                address:
                  type: string
                  minLength: 32
                  maxLength: 64
                minutes:
                  type: integer
                  minimum: 5
                  maximum: 1440
              required:
                - network
                - address
                - minutes
      parameters:
        - &id001
          in: header
          name: Origin
          required: true
          schema:
            type: string
            format: uri
          description: Exact application HTTPS origin
        - &id002
          in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
            pattern: ^[a-f0-9]{64}$
  /v1/auth/control/invitations/{id}/revoke:
    post:
      summary: Revoke a wallet invitation immediately
      description:
        Administrator only. Existing guest sessions stop authorizing on their
        next check.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties: {}
              required: []
      parameters:
        - *id001
        - *id002
        - in: path
          name: id
          required: true
          schema: &id003
            type: string
            format: uuid
  /v1/auth/control/actions:
    post:
      summary: Request a simulated deployment
      description:
        Any authorized session. Payload is immutable. Expires after at most 30
        minutes, bounded by the requesting session. Only the original session
        can execute.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Action'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentPayload'
      parameters:
        - *id001
        - *id002
  /v1/auth/control/actions/{id}/challenge:
    post:
      summary: Issue a fresh administrator action proof
      description:
        Requires admin. Bound to action, payload hash, signing wallet and
        session. Five-minute maximum, capped by request and administrator
        session. Replaces the same session/action challenge.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionChallenge'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                chainId:
                  type: string
                  minLength: 1
                  maxLength: 32
              required:
                - chainId
      parameters:
        - *id001
        - *id002
        - in: path
          name: id
          required: true
          schema: *id003
  /v1/auth/control/actions/{id}/approve:
    post:
      summary: Verify and consume an exact signed action proof
      description:
        Requires the issuing administrator session. Invalid matched proofs
        consume their nonce. Approval expires at the original challenge
        deadline; no new five-minute window is added.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Action'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                nonce:
                  type: string
                  pattern: ^[a-f0-9]{32}$
                message:
                  type: string
                  minLength: 1
                  maxLength: 4096
                signature:
                  type: string
                  minLength: 1
                  maxLength: 256
              required:
                - nonce
                - message
                - signature
      parameters:
        - *id001
        - *id002
        - in: path
          name: id
          required: true
          schema: *id003
  /v1/auth/control/actions/{id}/execute:
    post:
      summary: Commit a simulated deployment exactly once per action
      description:
        Original requesting session only. Both requester and administrator
        sessions must still authorize. Effect, consumption and audit share one
        SQLite transaction. No external deployment runs. A repeat returns 409.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties: {}
              required: []
      parameters:
        - *id001
        - *id002
        - in: path
          name: id
          required: true
          schema: *id003
  /v1/auth/control/actions/{id}/cancel:
    post:
      summary: Cancel a pending or approved action
      description:
        Original requesting session or administrator in the same application.
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid session or proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role, origin or CSRF denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Operation not found in this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description:
            Operation unavailable, already consumed, expired or conflicting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate or pending-state limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Policy or storage unavailable; deny access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties: {}
              required: []
      parameters:
        - *id001
        - *id002
        - in: path
          name: id
          required: true
          schema: *id003
components:
  securitySchemes:
    LoginContext:
      type: apiKey
      in: cookie
      name: __Host-gozne-login
      description: Secure; HttpOnly; SameSite=Strict; Path=/; no Domain
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-gozne-session
      description: Secure; HttpOnly; SameSite=Strict; Path=/; no Domain
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
          required:
            - code
            - message
            - request_id
      required:
        - error
    DeploymentPayload:
      type: object
      additionalProperties: false
      properties:
        project:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
        version:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
        environment:
          type: string
          enum:
            - preview
            - staging
            - production
      required:
        - project
        - version
        - environment
    Action:
      type: object
      additionalProperties: false
      properties:
        id: *id003
        application:
          type: string
        requester:
          type: string
        payload:
          $ref: '#/components/schemas/DeploymentPayload'
        payloadHash:
          type: string
          pattern: ^[a-f0-9]{64}$
        createdAt:
          type: integer
        expiresAt:
          type: integer
        status:
          type: string
          enum:
            - pending
            - approved
            - executed
            - canceled
            - expired
        approvedBy:
          type:
            - string
            - 'null'
        approvedAt:
          type:
            - integer
            - 'null'
        approvalExpiresAt:
          type:
            - integer
            - 'null'
        executedAt:
          type:
            - integer
            - 'null'
      required:
        - id
        - application
        - requester
        - payload
        - payloadHash
        - createdAt
        - expiresAt
        - status
        - approvedBy
        - approvedAt
        - approvalExpiresAt
        - executedAt
    Invitation:
      type: object
      additionalProperties: false
      properties:
        id: *id003
        network:
          type: string
          enum:
            - evm
            - solana
        address:
          type: string
        createdBy:
          type: string
        createdAt:
          type: integer
        expiresAt:
          type: integer
        revokedAt:
          type:
            - integer
            - 'null'
        acceptedAt:
          type:
            - integer
            - 'null'
      required:
        - id
        - network
        - address
        - createdBy
        - createdAt
        - expiresAt
        - revokedAt
        - acceptedAt
    InvitationCreated:
      type: object
      additionalProperties: false
      properties:
        id: *id003
        network:
          type: string
          enum:
            - evm
            - solana
        address:
          type: string
        expiresAt:
          type: integer
        roles:
          type: array
          items:
            const: reader
          minItems: 1
          maxItems: 1
        url:
          type: string
          format: uri
          description: Normal login URL; no secret bearer token
      required:
        - id
        - network
        - address
        - expiresAt
        - roles
        - url
    Deployment:
      type: object
      additionalProperties: false
      properties:
        actionId: *id003
        project: &id004
          type: string
        version: &id005
          type: string
        environment: &id006
          type: string
          enum:
            - preview
            - staging
            - production
        executedAt: &id007
          type: integer
      required:
        - actionId
        - project
        - version
        - environment
        - executedAt
    ExecutionReceipt:
      type: object
      additionalProperties: false
      properties:
        actionId: *id003
        project: *id004
        version: *id005
        environment: *id006
        executedAt: *id007
        simulated:
          type: boolean
          const: true
      required:
        - actionId
        - project
        - version
        - environment
        - executedAt
        - simulated
    Execution:
      type: object
      additionalProperties: false
      properties:
        action:
          $ref: '#/components/schemas/Action'
        receipt:
          $ref: '#/components/schemas/ExecutionReceipt'
      required:
        - action
        - receipt
    SessionMetadata:
      type: object
      additionalProperties: false
      properties:
        id: *id003
        identity:
          type: string
        createdAt:
          type: integer
        expiresAt:
          type: integer
      required:
        - id
        - identity
        - createdAt
        - expiresAt
    ControlOverview:
      type: object
      additionalProperties: false
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionMetadata'
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/Deployment'
      required:
        - actions
        - invitations
        - sessions
        - deployments
    ActionChallenge:
      type: object
      additionalProperties: false
      properties:
        nonce:
          type: string
          pattern: ^[a-f0-9]{32}$
        message:
          type: string
        expiresAt:
          type: integer
        signInInput:
          type: object
          additionalProperties: true
      required:
        - nonce
        - message
        - expiresAt
    Ok:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
      required:
        - ok